Sua co che event moi cho image service

+++ backend/services/image-service/image/src/main/java/vn/uit/edu/msshop/image/adapter/in/web/ImageEventListener.java
@@ -1,14 +1,18 @@
 package vn.uit.edu.msshop.image.adapter.in.web;
 
 import java.io.IOException;
+import java.time.Instant;
 
 import org.springframework.kafka.annotation.KafkaHandler;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import vn.uit.edu.msshop.image.adapter.in.web.mapper.ImageWebMapper;
+import vn.uit.edu.msshop.image.adapter.out.event.EventDocument;
+import vn.uit.edu.msshop.image.adapter.out.event.EventDocumentRepository;
 import vn.uit.edu.msshop.image.application.port.in.DeleteImageUseCase;
 import vn.uit.edu.msshop.image.application.port.in.RemoveImageFolderUseCase;
 import vn.uit.edu.msshop.image.domain.event.DeleteOldImageEvent;
@@ -22,17 +26,28 @@ public class ImageEventListener {
     private final DeleteImageUseCase deleteUseCase;
     private final RemoveImageFolderUseCase removeUseCase;
     private final ImageWebMapper mapper;
+    private final EventDocumentRepository eventDocumentRepo;
     
     @KafkaHandler
+    @Transactional
     public void handleDeleteOldImage(DeleteOldImageEvent event) {
-        final var command = mapper.toCommand(event.oldImagePublicId());
-        deleteUseCase.deleteImage(command);
+        if(!eventDocumentRepo.existsById(event.eventId())) {
+            final var command = mapper.toCommand(event.oldImagePublicId());
+            deleteUseCase.deleteImage(command);
+            eventDocumentRepo.save(EventDocument.builder().eventId(event.eventId()).receiveAt(Instant.now()).build());
+        }
+        
     }
 
     @KafkaHandler
+    @Transactional
     public void handleRollback(RollbackImageEvent event) throws IOException {
-        final var command = mapper.toCommand(event);
-        removeUseCase.rollbackImageFolder(command);
+        if(!eventDocumentRepo.existsById(event.eventId())) {
+            final var command = mapper.toCommand(event);
+            removeUseCase.rollbackImageFolder(command);
+            eventDocumentRepo.save(EventDocument.builder().eventId(event.eventId()).receiveAt(Instant.now()).build());
+        }
+        
     }
